gd32v/f103: add a Generic USB Display example
This is an example that illustrates support for RISC-V based GD32VF103 MCU and
its differences from STM32F103 which it resembles to an extent.

It drives a SSD1306 OLED display attached to the SPI bus and exposes it via USB
using the Generic USB Display [1] protocol.

[1] https://lore.kernel.org/dri-devel/20200529175643.46094-1-noralf@tronnes.org/

It also listens on a serial port, echoes the characters and toggles a LED on
each character, to demonstrate that interrupt-driven I/O works.

The example can also be built for a STM32F103. Details about the connection are
in config.h.